[REFACTOR] Move admin related APIs to adminController - CorkageInfo: list, add - CorkageStore: list, add
